+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Upgrade the `au.com.stocksoftware.idea.codestyle` artifact to version `1.17`.
 * Remove infrastructure supporting integration with the historic GWT "events" packages.
+* Migrate from using Elemental2 to Akasha when interacting with the Browser API.
 
 ### [v0.9](https://github.com/realityforge/gwt-keycloak/tree/v0.9) (2021-01-12) · [Full Changelog](https://github.com/realityforge/gwt-keycloak/compare/v0.8...v0.9)